+++ devices/stm32f429xx/stm32f429xx_init.c
@@ -60,7 +60,7 @@ void mriStm32f429xx_Init(Token* pParameterTokens)
 
 
 
-uint32_t Platform_GetDeviceMemoryMapXmlSize(void)
+size_t Platform_GetDeviceMemoryMapXmlSize(void)
 {
     return sizeof(g_memoryMapXml) - 1;
 }
